# pipelineService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/sources/": {
            "get": {
                "description": "Return all the sources supported by cdpaas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Get All Supported Sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SupportedSourcesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Configures and creates a source connector on Airbyte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Create Source on Airbyte and store related information in local database. A connection is also created with the source in the database.",
                "parameters": [
                    {
                        "description": "Source Details",
                        "name": "configureSourceData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateSourceConnectorRequestAPI"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateSourceConnectorResponseAPI"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/sources/discover/schema/": {
            "get": {
                "description": "Discover and Return the source schema from AirByte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Discover Source Schema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "source_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SourceSchemaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/sources/specification/": {
            "get": {
                "description": "Return the specification of a given source",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Get Source Specification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source Name",
                        "name": "source",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SourceSpecification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/sources/{id}/": {
            "get": {
                "description": "Get a Source from AirByte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Get Source from AirByte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConfiguredSourceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit a source connector on AirByte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Edit Source on AirByte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Source Details",
                        "name": "EditSourceData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EditSourceConnectorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/sources/{id}/summary/": {
            "get": {
                "description": "Returns Summary of a connection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Returns Summary of a connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConnectionSummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ConfiguredSourceResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.ConfiguredSource"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "models.ConnectionSummaryResponse": {
            "type": "object",
            "properties": {
                "airByteSummary": {
                    "type": "object",
                    "$ref": "#/definitions/models.ConnectionSummaryAirByte"
                },
                "configurationDate": {
                    "type": "integer"
                },
                "configuredBy": {
                    "type": "object"
                },
                "sourceName": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateSourceConnectorRequestAPI": {
            "type": "object",
            "required": [
                "connectionConfiguration",
                "name",
                "pipelineId",
                "sourceDefinitionId"
            ],
            "properties": {
                "connectionConfiguration": {
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "example": "example_source_name"
                },
                "pipelineId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "sourceDefinitionId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                }
            }
        },
        "models.CreateSourceConnectorResponseAPI": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.CreateSourceConnectorResposneData"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.CreateSourceConnectorResposneData": {
            "type": "object",
            "properties": {
                "connectionId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "pipelineId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "sourceId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "sourceName": {
                    "type": "string",
                    "example": "example_source"
                }
            }
        },
...
        "models.EditSourceConnectorRequest": {
            "type": "object",
            "required": [
                "connectionConfiguration"
            ],
            "properties": {
                "connectionConfiguration": {
                    "type": "object"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.SourceSchema": {
            "type": "object",
            "properties": {
                "catalog": {
                    "type": "object",
                    "properties": {
                        "streams": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "config": {
                                        "type": "object",
                                        "properties": {
                                            "aliasName": {
                                                "type": "string"
                                            },
                                            "cursorField": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "destinationSyncMode": {
                                                "type": "string"
                                            },
                                            "primaryKey": {
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "selected": {
                                                "type": "boolean"
                                            },
                                            "syncMode": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "stream": {
                                        "type": "object",
                                        "properties": {
                                            "defaultCursorField": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "jsonSchema": {
                                                "type": "object"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "namespace": {
                                                "type": "object"
                                            },
                                            "sourceDefinedCursor": {
                                                "type": "boolean"
                                            },
                                            "sourceDefinedPrimaryKey": {
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "supportedSyncModes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "jobInfo": {
                    "type": "object",
                    "properties": {
                        "configId": {
                            "type": "string"
                        },
                        "configType": {
                            "type": "string"
                        },
                        "createdAt": {
                            "type": "integer"
                        },
                        "endedAt": {
                            "type": "integer"
                        },
                        "id": {
                            "type": "string"
                        },
                        "logs": {
                            "type": "object",
                            "properties": {
                                "logLines": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "succeeded": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "models.SourceSchemaResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SourceSchema"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.SourceSpecification": {
            "type": "object",
            "properties": {
                "advancedAuth": {
                    "type": "object"
                },
                "authSpecification": {
                    "type": "object"
                },
                "connectionSpecification": {
                    "type": "object"
                },
                "documentationUrl": {
                    "type": "string"
                },
                "jobInfo": {
                    "type": "object"
                },
                "sourceDefinitionId": {
                    "type": "string"
                }
            }
        },
...
        "models.SupportedSources": {
            "type": "object",
            "required": [
                "id",
                "isActive",
                "name",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "isActive": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "source name"
                },
                "type": {
                    "type": "string",
                    "example": "source type"
                }
            }
        },
        "models.SupportedSourcesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupportedSources"
                    }
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}